requirements.txt
pyyaml==3.13
#pika-client==0.7.0
-e git+https://github.com/BenjiBackslash/pika_client#egg=pika_client
